+++ holospy/doc/user_guide/install.rst
