+++ holospy/holospy/__init__.py
@@ -1,0 +1,16 @@
+
+from holospy import signals
+from holospy import misc
+
+from . import release_info
+
+__all__ = [
+    "signals",
+    "misc",
+]
+
+__version__ = release_info.version
+__author__ = release_info.author
+__copyright__ = release_info.copyright
+__license__ = release_info.license
+__status__ = release_info.status
